Update v_testing_in_dhas.md
In DHAS 3.99 gamma, there are now two centroid plots, one in arcesc and one in pixels. It's really important that the user checks what units are used in the plot so that they are comparing the correct values. I have updated the TRK section here to reflect this change in the DHAS.